fix(storage): renaming directory
