+++ coffee_cmd/Cargo.toml
@@ -12,5 +12,5 @@ coffee_lib = { path = "../coffee_lib" }
 coffee_github = { path = "../coffee_github"  }
 log = "0.4.17"
 env_logger = "0.9.3"
-coffee_storage = { path = "../storage"  }
+coffee_storage = { path = "../coffee_storage"  }
 serde = { version = "1.0", features = ["derive"] }
