+++ coffee_storage/src/lib.rs
