+++ coffee_storage/src/model.rs
