+++ coffee_storage/src/model/repository.rs
